\name{get.adj.daily.close}
\alias{get.adj.daily.close}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.adj.daily.close(dates, seccodes, per.seccode = 1)
}
\arguments{
  \item{dates}{Set of dates between which you need the
  results}

  \item{seccodes}{Set of codes for which result is needed}

  \item{per.seccode}{Set if needed for a particular
  seccode}
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah
}

