\name{get.coas.and.descriptions}
\alias{get.coas.and.descriptions}
\title{Get Coas And Descriptions)}
\author{Ian MacGillivray}
\description{
Returns all available COAs and their descriptions.
}
\usage{
get.coas.and.descriptions()
}
\examples{
	\dontrun{
		get.coas.and.descriptions()
	}
}