\name{get.currency.rates}
\alias{get.currency.rates}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.currency.rates(start.date = "2011-01-01", end.date = "2011-02-01",
  from.curr = "all", to.curr = "USD", rate.type = "SPOT")
}
\arguments{
  \item{start.date}{start.date in Y-m-d format}

  \item{end.date}{end.date in Y-m-d format}

  \item{from.curr}{set of currencies to be converted}

  \item{to.curr}{which currency to be converted to}

  \item{rate.type}{SPOT or which forward contract etc}
}
\value{
data.frame
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah
}

