\name{get.description.for.coa}
\alias{get.description.for.coa}
\title{Get Description For Coa)}
\author{Ian MacGillivray}
\description{
Returns the description for a given goa.
}
\usage{
get.description.for.coa(coa)
}
\arguments{
\item{coa}{A QAD COA.}
}
\examples{
	\dontrun{
		get.description.for.coa("ADPL")
	}
}
