\name{get.dividends.payable}
\alias{get.dividends.payable}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.dividends.payable(dates, seccodes, quarterly = TRUE, most.recent = TRUE)
}
\arguments{
  \item{dates}{dates for which result is needed}

  \item{seccodes}{seccodes for which result is needed}

  \item{quarterly}{boolean to set whether quarterly or
  annual returns are needed}

  \item{most.recent}{boolean}
}
\value{
matrix
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Ian MacGillivray
}

