\name{get.earnings.surprise}
\alias{get.earnings.surprise}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.earnings.surprise(dates, seccodes)
}
\arguments{
  \item{dates}{dates for which earnings surprise is needed}

  \item{seccodes}{seccodes for which earnings surprise is
  needed}
}
\value{
matrix with missing vals as NA
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah
}

