\name{get.employees}
\alias{get.employees}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.employees(dates, seccodes, quarterly = TRUE, most.recent = TRUE)
}
\arguments{
  \item{dates}{dates for which employees number are needed}

  \item{seccodes}{seccodes for which employees number are
  needed}

  \item{quarterly}{boolean to set quarterly or not}

  \item{most.recent}{if most recent is needed or not}
}
\value{
matrix
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Ian MacGillivray
}

