\name{get.eps}
\alias{get.eps}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.eps(dates, seccodes)
}
\arguments{
  \item{dates}{dates for which eps is needed}

  \item{seccodes}{seccodes for which eps is needed}
}
\value{
array
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah
}

