\name{get.holidays}
\alias{get.holidays}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.holidays(start.date = "2009-01-01", end.date = "2012-02-01")
}
\arguments{
  \item{start.date}{start.date in Y-m-d format}

  \item{end.date}{end.date in Y-m-d format}
}
\value{
Date array of holidays in Y-m-d format
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah
}

