\name{get.licensed.tables}
\alias{get.licensed.tables}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
Get the list of tables permissioned}
\usage{
get.licensed.tables(dbName = "QAI")
}
\arguments{
  \item{dbName}{database name string}
}
\value{
tables permissioned
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved. Get
the list of tables permissioned
}
\author{
Sameena Shah
}

