\name{get.market.cap}
\alias{get.market.cap}
\title{This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.market.cap(dates, seccodes)
}
\arguments{
  \item{dates}{dates}

  \item{seccodes}{seccodes}
}
\value{
matrix
}
\description{
This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah
}

