\name{get.qad.connection}
\alias{get.qad.connection}
\title{Get QAD Connection}
\author{Ian MacGillivray}
\description{
Returns a connection to a QAD database, used extensively throughout the library. Before this method is first used, you must to alter the connection function using \code{\link{set.qad.connection}}.
}
\usage{
get.qad.connection()
}
\seealso{
	\code{\link{set.qad.connection}}
}
\examples{
	\dontrun{
		get.qad.connection()
	}
}