\name{get.tangible.book.value.per.share}
\alias{get.tangible.book.value.per.share}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
Returns the tangible book value per share for a set of dates and seccodes}
\usage{
get.tangible.book.value.per.share(dates, seccodes, quarterly = TRUE,
  most.recent = TRUE)
}
\arguments{
  \item{dates}{dates}

  \item{seccodes}{seccodes}

  \item{quarterly}{boolean}

  \item{most.recent}{boolean}
}
\value{
matrix
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
Returns the tangible book value per share for a set of
dates and seccodes
}
\author{
Ian MacGillivray
}

