\name{get.unadj.daily.close}
\alias{get.unadj.daily.close}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.unadj.daily.close(dates, seccodes, per.seccode = 1)
}
\arguments{
  \item{dates}{dates}

  \item{seccodes}{seccodes}

  \item{per.seccode}{per.seccode}
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}

