\name{get.underwriting.profit}
\alias{get.underwriting.profit}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.underwriting.profit(dates, seccodes, quarterly = TRUE,
  most.recent = TRUE)
}
\arguments{
  \item{dates}{dates}

  \item{seccodes}{seccodes}

  \item{quarterly}{if only quarterly}

  \item{most.recent}{if only most recent}
}
\value{
matrix
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}

