\name{set.qad.connection}
\alias{set.qad.connection}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
Sets user connection}
\usage{
set.qad.connection(func, ...)
}
\arguments{
  \item{func}{Pass the function that you need to use to
  establish connection}

  \item{...}{any additional arguments that you want to pass
  for func}
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved. Sets
user connection
}
\author{
Sameena Shah
}

