\name{example.connection}
\alias{example.connection}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
example.connection(server, uid, pwd, other = "")
}
\arguments{
  \item{server}{server address}

  \item{uid}{userid}

  \item{pwd}{password}

  \item{other}{any other information like database name
  etc, none by default}
}
\value{
returns an odbcConnection object
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah
}

