\name{get.current.betas}
\alias{get.current.betas}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.current.betas(seccodes)
}
\arguments{
  \item{seccodes}{seccodes}
}
\value{
array with NA where no data obtained for the seccode
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}

