\name{get.current.ratio}
\alias{get.current.ratio}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.current.ratio(dates, seccodes, quarterly)
}
\arguments{
  \item{dates}{dates}

  \item{seccodes}{seccodes}

  \item{quarterly}{if quarterly is needed}
}
\value{
matrix
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Ian MacGillivray
}

