\name{get.description.for.coa}
\alias{get.description.for.coa}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved}
\usage{
get.description.for.coa(coa)
}
\arguments{
  \item{coa}{coa code for which descrition is needed}
}
\value{
matrix
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved
}
\author{
Ian MacGillivray
}

