\name{get.info.from.rkd}
\alias{get.info.from.rkd}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.info.from.rkd(dates, seccodes, type, quarterly = TRUE,
  most.recent = TRUE)
}
\arguments{
  \item{dates}{dates}

  \item{seccodes}{seccodes}

  \item{type}{type}

  \item{quarterly}{quarterly}

  \item{most.recent}{most.recent}
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah
}

