\name{get.total.liabilities}
\alias{get.total.liabilities}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.total.liabilities(dates, seccodes, quarterly = TRUE, most.recent = TRUE)
}
\arguments{
  \item{dates}{dates}

  \item{seccodes}{seccodes}

  \item{quarterly}{boolean}

  \item{most.recent}{boolean}
}
\value{
matrix
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Ian MacGillivray
}

