\name{get.trade.dates}
\alias{get.trade.dates}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.}
\usage{
get.trade.dates(dates)
}
\arguments{
  \item{dates}{dates}
}
\value{
vector of trade.dates Takes a vector of dates and returns a
vector of valid trading dates for information arriving on
these dates. For extra precision, first call
add.open.close.date.indexes and choose a date based on the
period (e.g. close-close) you care about.
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
}
\author{
Sameena Shah and Ian MacGillivray
}

