\name{get.trading.dates.from.db}
\alias{get.trading.dates.from.db}
\title{#This source code is provided under the BSD license and is provided AS IS with no warranty or guarantee of fit for purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
Assumes that the days on which some of the largest mcap companies didnt trade is not a valid trading day}
\usage{
get.trading.dates.from.db(start.date = "2011-01-01",
  end.date = "2011-02-01")
}
\arguments{
  \item{start.date}{start.date in Y-m-d format}

  \item{end.date}{end.date in Y-m-d format}
}
\value{
trade.dates vector
}
\description{
#This source code is provided under the BSD license and is
provided AS IS with no warranty or guarantee of fit for
purpose.  See the project's LICENSE.txt for details.
#Copyright Thomson Reuters 2013. All rights reserved.
Assumes that the days on which some of the largest mcap
companies didnt trade is not a valid trading day
}
\author{
Ian MacGillivray and Sameena Shah
}

